+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - various UI/UX improvements
+- cdk deployment for us-gov-east-1
+- cloud formation template for us-gov-west-1
 
 ## [1.0.1] - 2023-06-30